fix(policy): move doctor shared types to leaf module
